# algohive/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  groups.CreateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
      scope_id:
        type: string
    required:
    - name
    - scope_id
    type: object
  groups.UpdateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  models.Catalog:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
    type: object
  models.Competition:
    properties:
      api_environment:
        $ref: '#/definitions/models.Catalog'
      api_environment_id:
        type: string
      api_theme:
        type: string
      description:
        type: string
      finished:
        type: boolean
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      show:
        type: boolean
      title:
        type: string
      tries:
        items:
          $ref: '#/definitions/models.Try'
        type: array
    type: object
  models.Group:
    properties:
      competitions:
        items:
          $ref: '#/definitions/models.Competition'
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scope_id:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Role:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        type: integer
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Scope:
    properties:
      catalogs:
        items:
          $ref: '#/definitions/models.Catalog'
        type: array
      description:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  models.Try:
    properties:
      attempts:
        type: integer
      competition_id:
        type: string
      end_time:
        type: string
      id:
        type: string
      puzzle_id:
        type: string
      puzzle_index:
        type: integer
      puzzle_lvl:
        type: string
      score:
        type: number
      start_time:
        type: string
      step:
        type: integer
      user_id:
        type: string
    type: object
  models.User:
    properties:
      blocked:
        type: boolean
      email:
        type: string
      firstname:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      last_connected:
        type: string
      lastname:
        type: string
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  roles.CreateRoleRequest:
    properties:
      name:
        type: string
      permission:
        type: integer
      scopes_ids:
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
info:
  contact:
    email: ericphlpp@proton.me
    name: AlgoHive Support
  description: This is the API documentation for the AlgoHive API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger AlgoHive API
  version: 1.0.0
paths:
...
  /groups:
    get:
      consumes:
      - application/json
      description: Get all groups, only accessible to users with the GROUPS permission
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all groups
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Create a group, only accessible to users with the GROUPS permission
      parameters:
      - description: Group to create
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/groups.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a group
      tags:
      - Groups
  /groups/{group_id}:
    delete:
      consumes:
      - application/json
      description: Delete a group and cascade delete all users and competitions associated
        with the group
      parameters:
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a group and cascade delete all users and competitions associated
        with the group
      tags:
      - Groups
    get:
      consumes:
      - application/json
      description: Get a group
      parameters:
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a group
      tags:
      - Groups
    put:
      consumes:
      - application/json
      description: Update a group name and description
      parameters:
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      - description: Group to update
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/groups.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a group name and description
      tags:
      - Groups
...
  /groups/scope/{scope_id}:
    get:
      consumes:
      - application/json
      description: Get all the groups from a given scope and list the users in each
        group
      parameters:
      - description: Scope ID
        in: path
        name: scope_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all the groups from a given scope
      tags:
      - Groups
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get all Roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      security:
      - Bearer: []
      summary: Get all Roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new Role
      parameters:
      - description: Role Profile
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/roles.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new Role
      tags:
      - Roles
  /roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: delete a role and cascade first to roles_scopes and user_roles
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: delete a role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get a Role by ID
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a Role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update a Role by ID
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Role Profile
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a Role by ID
      tags:
      - Roles
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"